orca: minor cleanups
